Add base classes package and update imports
