+++ organizmy/__init__.py
@@ -1,0 +1,5 @@
+from .organizm import Organizm
+from .zwierze import Zwierze
+from .roslina import Roslina
+
+__all__ = ["Organizm", "Zwierze", "Roslina"]
